Refactor app initialization:
The former loader.js provided app initialization and theme overrides of all modules. Replace the initApp() function by using the standard JS module pattern of a function expression that runs on load.  This simplifies the loading dramatically.  For custom app initialization support, we will fold that into app.js by providing custom modules on a per-theme basis (TODO: to be discussed). Move the preventCompatibilityIssues() polyfill into app.js and execute it after the MetacatUI object has been created.  Do the same for getMapKey() and getMetacatContext(), however I think we should move those two functions into the the MapModel and Theme objects respectively (TODO: to be discussed). Ultimately, it's cleaner to not have these configuration options in index.html (data-map-key and data-metacat-context).

Configure RequireJS the way have have done normally, providing paths to modules and shims for backward compatibility. Don't set the RequireJS "map" attribute until after the theme is loaded.

Support theme overrides by introducing a new Theme backbone model, and requiring it in app.js. Theme is responsible for all things theme, and has a static getThemeName() method that extracts the index.html data-theme attribute value.  This is then used in a local scope to build a path to the configuration file for the named theme.  I migrated the theme format to plain JSON, and we new up a Theme instance with the JSON theme object used as the attributes of the Theme object so we can get() and set() theme properties easily. Once the theme is loaded, it is stored in the MetacatUI.theme attribute.

With the theme in place, support module overrides the same was as before where we add the RequireJS configuration's "map" attribute with the MetacatUI.theme.get("moduleMap") object value.  Template, JS, and CSS overrides work the same as before.

Application views rely on both MetacatUI and MetacatUI.theme to be available, so proceed with the rest of the app loading in the callback to requiring the theme configuration. This avoids any race conditions.

The main advantage to this refactor is that all app instantiation is done in app.js and is therefore usable in the unit test framework. We also get away from injecting <script> tags into the DOM <head> and then triggering events to load the scripts. The auto-executing module pattern seems to be a cleaner route, and we leave the module injection to RequireJS.

refs #509