Migrate theme format to JSON:
Copy each theme's config.js file to config.json, and reformat the JSON
files to be JSON -compliant (no JS code or comments).
Remove config.js files for themes that have no JS code needed to
customize the app. For themes with custom JS code, modify the config.js
files to remove the static config lines (now in the JSON), and keep the
JS code.  We will plan in a standard  app customization pattern and call
this code with that mechanism (TODO).

refs #509