Add draft saving support
Use the LocalForage library to handle saving models and collections 
offline to IndexedDB, WebSQL, or localStorage, depending on the age and 
capabilities of the browser. The objects are stored as key/value pairs, 
with the object id as the key, and the object itself as the value.

Modify DataPackage.initialize() to create an instance of LocalForage as 
the "drafts" attribute.
Add the saveDraft() method to save drafts offline.
Add the handleSavedDraft() method as a callback to the saving event. For 
now it just logs what happened.
 
refs #216