Add a NestedModel to bubble events
The NestedModel object overrides the Backbone.Model set() method in order to trigger change events on the parent if there are change events coming from nested Backbone.Model values for any attribute of the parent model. After bubbling the change event up to the parent, it then continues to call the super() Backbone.Model.set() method with the passed in arguments.

refs #216