Set listeners on arrayed backbone models
In order to catch events from sub-models, we also need to listen on the 
sub-models that are added to arrays of Backbone.Model instances.
Also, don't listen to events for parentModel, dataONEObject, and 
personnel attributes to avoid circular logic and stack overflow 
exceptions. This needs some refining - we need to track down the 
circular logic - it is related to calls to `createID()` for parties and 
creating `project/personnel` entries.

refs #216