Don't bypass model.set()
There are places in the EML211.serialize() method that would get a 
reference to an existing array previously set in the model, and then 
would push() items into that array.  This bypasses the model.set() call, 
and so change events don't get generated. Copy the array, modify the 
copy, and call model.set() instead.

refs #216